backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Machine Learning - Using alternative that doesn't require compilation
# fasttext-wheel>=0.9.2  # Commented out due to compilation issues
# Alternative: Use gensim or scikit-learn for basic language detection
scikit-learn>=1.3.0
numpy>=1.24.0

# Rate limiting and caching
slowapi>=0.1.9
redis>=5.0.0

# Configuration and utilities
python-dotenv>=1.0.0

# Optional: For production deployment
# gunicorn>=21.2.0
# supervisor>=4.2.5
